Move neighborhoods to different file
